Rename arrow.test to arrow.fx.test in arrow-fx-test module
