+++ arrow-fx-reactor/src/test/kotlin/arrow/fx/MonoKTest.kt
@@ -19,8 +19,8 @@ import arrow.fx.reactor.value
 import arrow.fx.typeclasses.ExitCase
 import arrow.core.test.UnitSpec
 import arrow.core.test.generators.GenK
-import arrow.test.laws.AsyncLaws
-import arrow.test.laws.TimerLaws
+import arrow.fx.test.laws.AsyncLaws
+import arrow.fx.test.laws.TimerLaws
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import io.kotlintest.matchers.startWith
