+++ arrow-fx-test/src/main/kotlin/arrow/fx/test/eq/EqK.kt
@@ -1,4 +1,4 @@
-package arrow.test.eq
+package arrow.fx.test.eq
 
 import arrow.Kind
 import arrow.core.extensions.either.eq.eq
