+++ arrow-fx-test/src/main/kotlin/arrow/fx/test/laws/ConcurrentLaws.kt
@@ -1,4 +1,4 @@
-package arrow.test.laws
+package arrow.fx.test.laws
 
 import arrow.Kind
 import arrow.core.Left
