+++ arrow-fx-test/src/main/kotlin/arrow/fx/test/laws/Law.kt
@@ -1,4 +1,4 @@
-package arrow.test.laws
+package arrow.fx.test.laws
 
 import arrow.core.test.generators.tuple2
 import arrow.core.test.generators.tuple3
