+++ arrow-fx-test/src/main/kotlin/arrow/fx/test/laws/MonadThrowLaws.kt
@@ -1,4 +1,4 @@
-package arrow.test.laws
+package arrow.fx.test.laws
 
 import arrow.Kind
 import arrow.core.extensions.eq
@@ -7,7 +7,7 @@ import arrow.core.test.generators.fatalThrowable
 import arrow.core.test.generators.throwable
 import arrow.core.test.laws.Law
 import arrow.core.test.laws.MonadErrorLaws
-import arrow.test.eq.throwableEq
+import arrow.fx.test.eq.throwableEq
 import arrow.typeclasses.Apply
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
