+++ arrow-fx/src/test/kotlin/arrow/fx/FiberTest.kt
@@ -10,9 +10,9 @@ import arrow.fx.extensions.io.applicative.applicative
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.monoid
 import arrow.fx.typeclasses.Fiber
-import arrow.test.eq.eq
-import arrow.test.eq.eqK
-import arrow.test.generators.genK
+import arrow.fx.test.eq.eq
+import arrow.fx.test.eq.eqK
+import arrow.fx.test.generators.genK
 import io.kotlintest.properties.Gen
 
 class FiberTest : UnitSpec() {
