+++ arrow-fx/src/test/kotlin/arrow/fx/SemaphoreTest.kt
@@ -13,7 +13,7 @@ import arrow.fx.extensions.io.concurrent.parMapN
 import arrow.fx.extensions.io.functor.unit
 import arrow.fx.extensions.io.monad.flatMap
 import arrow.fx.extensions.io.monad.map
-import arrow.test.eq.eq
+import arrow.fx.test.eq.eq
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
 import kotlinx.coroutines.Dispatchers
